chore(multistream-select): replace `futures_ringbuf` with TCP
We seem to have a flaky test in `multistream-select`. I could reproduce it locally but not fully track down what the issue is. I think it is related to the "sockets" being dropped to early or something.

Once I swapped them out for a TCP listener, the problem did not occur again.

Pull-Request: #4693.