deps(identity): bump `ring` to `v0.17.5`
The `libp2p-identity` crate is a dependency of the entire ecosystem. We should keep its dependencies up to date as much as possible. At the moment, we can't bump `ring` across the entire workspace because we still need releases of `rustls` that update to the new `ring` version. Not all our users depend in `libp2p-tls` or `libp2p-quic` though. For those users, a bump in `libp2p-identity` will be useful.

Pull-Request: #4779.