fix(webrtc-utils): read buffer only if empty message
This PR adds a return of `Poll::Ready(Ok(0))` when the `message` vector is empty or not present.

This missing return was causing the stream not to finish, then to be dropped, causing the request response protocol to fail over WebRTC.

Now that empty or non existent message vectors get returned as `Ok(0)` the stream can be read and request response works now over WebRTC.

Pull-Request: #5439.